+++ src/components/home/HotCollections.jsx
@@ -1,13 +1,69 @@
-import React from "react";
+import React, {useState} from "react";
 import { Link } from "react-router-dom";
 import AuthorImage from "../../images/author_thumbnail.jpg";
 import nftImage from "../../images/nftImage.jpg";
-import axios from "axios";
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 
+
+
+
 const HotCollections = () => {
+  const ItemDetails = () => {
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, []);
+
+  const [itemDetails, setItemDetails] = useState([])
+
+  useEffect (() => {
+    async function fetchItemDetails() {
+      const {data} = await axios.get (
+        `https://us-central1-nft-cloud-functions.cloudfunctions.net/hotCollections`
+      )
+      setItemDetails(data)
+    }
+    fetchItemDetails()
+  }, [])
+
+  const settings = {
+    dots: true,
+    infinite: true,
+    speed: 500,
+    slidesToShow: 4,
+    slidesToScroll: 1,
+    responsive: [
+      {
+        breakpoint: 1024,
+        settings: {
+          slidesToShow: 3,
+          slidesToScroll: 1,
+          infinite: true,
+          dots: true,
+        },
+      },
+      {
+        breakpoint: 768,
+        settings: {
+          slidesToShow: 2,
+          slidesToScroll: 1,
+          dots: true,
+        },
+      },
+      {
+        breakpoint: 480,
+        settings: {
+          slidesToShow: 1,
+          slidesToScroll: 1,
+          dots: true,
+        },
+      },
+    ],
+  };
+
+ 
+
   return (
     <section id="section-collections" className="no-bottom">
       <div className="container">
@@ -18,29 +74,40 @@ const HotCollections = () => {
               <div className="small-border bg-color-2"></div>
             </div>
           </div>
-          {new Array(4).fill(0).map((_, index) => (
-            <div className="col-lg-3 col-md-6 col-sm-6 col-xs-12" key={index}>
-              <div className="nft_coll">
-                <div className="nft_wrap">
-                  <Link to="/item-details">
-                    <img src={nftImage} className="lazy img-fluid" alt="" />
-                  </Link>
-                </div>
-                <div className="nft_coll_pp">
-                  <Link to="/author">
-                    <img className="lazy pp-coll" src={AuthorImage} alt="" />
-                  </Link>
-                  <i className="fa fa-check"></i>
-                </div>
-                <div className="nft_coll_info">
-                  <Link to="/explore">
-                    <h4>Pinky Ocean</h4>
-                  </Link>
-                  <span>ERC-192</span>
+          <Slider {...settings}>
+            {new Array(4).fill(0).map((_, index) => (
+              <div key={index} className="px-2">
+                <div className="nft_coll">
+                  <div className="nft_wrap">
+                    <Link to="/item-details">
+                      <img
+                        src={nftImage}
+                        className="lazy img-fluid"
+                        alt="NFT"
+                      />
+                    </Link>
+                  </div>
+                  <div className="nft_coll_pp">
+                    <Link to="/author">
+                      <img
+                        className="lazy pp-coll"
+                        src={AuthorImage}
+                        alt="Author"
+                      />
+                    </Link>
+                    <i className="fa fa-check"></i>
+                  </div>
+                  <div className="nft_coll_info">
+                    <Link to="/explore">
+                      <h4>Pinky Ocean</h4>
+                    </Link>
+                    <span>ERC-192</span>
+                  </div>
                 </div>
               </div>
-            </div>
-          ))}
+      
+            ))}
+          </Slider>
         </div>
       </div>
     </section>
